feat(pie-docs): DSW-000 centralise more information link for code pages
