+++ .changeset/friendly-cows-matter.md
@@ -1,0 +1,5 @@
+---
+"pie-docs": patch
+---
+
+[Changed] - New shortcode for a more info notification to use in each component code page
